chore: rename SMS fraud to SMS pumping
