+++ README.md
@@ -86,7 +86,7 @@ Integrate Fingerprint Bot Detection with your Web Application Firewall and dynam
 
 Prevent financial losses from SMS pumping. Link every verification text message to a browser fingerprint and limit the number of verification requests from a single browser.
 
-[📱 SMS Fraud Protection Live Demo](https://demo.fingerprint.com/sms-fraud)
+[📱 SMS Fraud Protection Live Demo](https://demo.fingerprint.com/sms-pumping)
 
 ## Documentation and Support
