+++ src/pages/api/admin/reset.ts
@@ -13,7 +13,7 @@ import { getAndValidateFingerprintResult } from '../../../server/checks';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { deleteBlockedIp } from '../../../server/botd-firewall/blockedIpsDatabase';
 import { syncFirewallRuleset } from '../../../server/botd-firewall/cloudflareApiHelper';
-import { SmsVerificationDatabaseModel } from '../../../server/sms-fraud/database';
+import { SmsVerificationDatabaseModel } from '../../../server/sms-pumping/database';
 
 export type ResetResponse = {
   message: string;
