+++ src/pages/api/sms-pumping/send-verification-sms.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { Severity } from '../../../server/checkResult';
 import { getAndValidateFingerprintResult } from '../../../server/checks';
 import { isValidPostRequest } from '../../../server/server';
-import { RealSmsPerVisitorModel, SmsVerificationDatabaseModel } from '../../../server/sms-fraud/database';
+import { RealSmsPerVisitorModel, SmsVerificationDatabaseModel } from '../../../server/sms-pumping/database';
 import { ONE_SECOND_MS, readableMilliseconds } from '../../../shared/timeUtils';
 import { Op } from 'sequelize';
 import { pluralize } from '../../../shared/utils';
@@ -14,7 +14,7 @@ import {
   REAL_SMS_LIMIT_PER_VISITOR,
   SMS_FRAUD_COPY,
   TEST_PHONE_NUMBER,
-} from '../../../server/sms-fraud/smsFraudConst';
+} from '../../../server/sms-pumping/smsPumpingConst';
 
 export type SendSMSPayload = {
   requestId: string;
@@ -164,7 +164,7 @@ export default async function sendVerificationSMS(req: NextApiRequest, res: Next
   try {
     await sendSms(
       phone,
-      `Your verification code for demo.fingerprint.com/sms-fraud is ${verificationCode}.`,
+      `Your verification code for demo.fingerprint.com/sms-pumping is ${verificationCode}.`,
       visitorId,
     );
